Just a tiny, crucial Infinity
The problem with floating-point is how it fills up its memory: you either have very large numbers or very precise numbers, not both.

When parsing of numbers with large negative exponents, the result may or may NOT be zero depending on what datatype is used. This results in some funky numbers with the BPM and also the actual timings used in-game.

Using `double` for this isn't a bad solution, but it would mess around with so much code that it's reserved for a later commit which needs more testing.

This commit makes Aspire 2017 *mostly* work.