Redo timing point parsing
Consolidate `beatLength` and `velocity` into a single variable, use the inherited flag, and deal with NaN values by setting default values for them.

This commit will break existing beatmap databases, because the inherited flag is stored inverted in the DB [0 for BPM, 1 for SV] instead of the file format spec [1 for BPM, 0 for SV], messing with every stored map's timing info.